+++ Sources/Core/ImageRequest.swift
@@ -102,6 +102,10 @@ public struct ImageRequest: CustomStringConvertible {
         /// )
         /// ```
         public static let imageIdKey: ImageRequest.UserInfoKey = "github.com/kean/nuke/imageId"
+
+        /// The image scale to be used. By default, the scale matches the scale
+        /// of the current display.
+        public static let scaleKey: ImageRequest.UserInfoKey = "github.com/kean/nuke/scale"
     }
 
     // MARK: Initializers